fixed (long int) error message

+++ src/news.c
@@ -558,8 +558,8 @@ int news_saveArticles( xmlTextWriterPtr writer )
          xmlw_attr(writer,"title","%s",ntitle);
          xmlw_attr(writer,"desc","%s",ndesc);
          xmlw_attr(writer,"faction","%s",article_ptr->faction);
-         xmlw_attr(writer,"date","%li",article_ptr->date);
-         xmlw_attr(writer,"date_to_rm","%li",article_ptr->date_to_rm);
+         xmlw_attr(writer,"date","%li",(long int) article_ptr->date);
+         xmlw_attr(writer,"date_to_rm","%li",(long int) article_ptr->date_to_rm);
          xmlw_attr(writer,"id","%i",article_ptr->id);
 
          if (article_ptr->tag!=NULL)
